updte .net versions and fix bugs ('.' in names being removed on preview attachments, crashes on email without body)

+++ src/XstReader.Exporter/MsgKit/Structures/Properties.cs
@@ -470,11 +470,11 @@ namespace XstReader.Exporter.MsgKit.Structures
                             break;
 
                         case TypeCode.SByte:
-                            data = BitConverter.GetBytes((sbyte)obj);
+                            data = BitConverter.GetBytes((short)(sbyte)obj);
                             break;
 
                         case TypeCode.Byte:
-                            data = BitConverter.GetBytes((byte)obj);
+                            data = BitConverter.GetBytes((ushort)(byte)obj);
                             break;
                         case TypeCode.Int16:
                             data = BitConverter.GetBytes((short)obj);
